refactor(akto): move aktoApi helpers into plugins/akto to align with gateway arch
Move src/utils/aktoApi.ts → plugins/akto/helpers.ts and its tests into
plugins/akto/helpers.test.ts so all Akto logic is self-contained under
the plugin directory. Update scan.ts and scan.test.ts imports accordingly.
